Make bgfill prompt item work in Fish
For whatever reason, using the Erase in Line sequence (CSI K) like we
were doing previously doesn't work properly in Fish. Now, we set the
background colour by hand (for Fish), then use EL to clear the line as
before (for Bash). This has to be a global $format modification so that
it comes after the $cmd_duration.